modified hospital
