+++ 01-python/infrastructure/road.py
@@ -1,5 +1,3 @@
 def draw_road():
-    for i in range(10):
-        print("=",end="")
-    print()
+    print("road not found")
     return
