+++ 01-python/safety/safety.py
@@ -4,6 +4,6 @@ from .policestation import draw_policestation
 
 def draw_safety():
     draw_firestation()
-    draw_hospital()
+    draw_hospital(13)
     draw_policestation()
     return
